requirements.txt
<<<<<<< HEAD
setuptools==41.2.0
tensorflow==2.3.0
matplotlib==3.1.1
gym==0.17.2
ghast==0.1.0
dgl==0.5.1
future==0.18.2
numpy==1.17.3
periodictable==1.5.1
tensorflow-gan==2.0.0
tensorflow-hub==0.9.0
imageio==2.9.0
tqdm==4.48.2
=======
absl-py==0.11.0
aniso8601==8.0.0
astunparse==1.6.3
attrs==20.3.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cheroot==6.6.0
click==7.1.2
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
dgl==0.5.1
dm-tree==0.1.5
Flask==1.1.2
flask-restplus==0.13.0
future==0.18.2
gast==0.3.3
ghast==0.1.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gym==0.17.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==8.6.0
networkx==2.5
numpy==1.17.3
oauthlib==3.1.0
opt-einsum==3.3.0
periodictable==1.5.1
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gan==2.0.0
tensorflow-hub==0.9.0
tensorflow-probability==0.11.1
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.48.2
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
>>>>>>> 6e1981d60bcb89308222b666db6a983230defd41
